Read the duration of the audio file to show in the card and rename the component folders to plural
