+++ app/components/shared/audioWaveButtons/AudioWaveButtonRippleComponent.js
