+++ app/components/shared/cardWithSoundWaves/CardWithSoundWaveInfoComponent.js
