+++ app/components/shared/cards/CardInfoComponent.js
