+++ app/views/home/HomeView.js
@@ -12,11 +12,11 @@ import yourStoryLong from '../../assets/audios/your_story_long.mp3';
 const HomeView = () => {
   const renderCards = () => {
     const items = [
-      { id: 1, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: saftyPlan, duration: 14 },
-      { id: 2, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: yourStory, duration: 5 },
-      { id: 3, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: yourStoryLong, duration: 60 },
-      { id: 4, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: null, duration: 0 },
-      { id: 5, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: null, duration: 0 },
+      { id: 1, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: saftyPlan },
+      { id: 2, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: yourStory },
+      { id: 3, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: yourStoryLong },
+      { id: 4, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: null },
+      { id: 5, title: 'បម្រែបម្រួលរាង្គកាយក្មេងប្រុសនៅពេលពេញវ័យ', description: 'ជំងឺនានាដែលអ្នកគួរយល់ដឹងពាក់ពន្ធផ្លូវភេទ និងវិធីសាស្រ្តសម្រាប់ទប់ស្កាត់ និង បង្ការ ជាមួយមតិយោបល់នានាពីអ្នកឯកទេស', audio: null },
     ]
 
     return <CardWithSoundWaveListComponent items={items} />
